+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.4.13
 * Added: Quick fix for incorrect extends and implements (#940)
 * Fixed: Resolving type from inline method calls (#868)
+* Fixed: Automatic import was placed after class Doc comment (#842)
 
 ## 1.4.12
 * Bugfix: "import class quickfix" missing in some cases (#1132)